delete function expanded with password check

+++ backend/controllers/user/deleteUser.js
@@ -1,9 +1,32 @@
 import { User } from "../../models/User.js";
 import { Conversation } from "../../models/Conversation.js";
+import { compare } from "../../utils/crypto.js";
 
 export const deleteUser = async (req, res, next) => {
   try {
-    const { userId } = req.body;
+    const { userId, password } = req.body;
+
+    //FIND USER
+    const user = await User.findOne({ _id: userId });
+
+    const enteredPassword = password;
+    const hashedPassword = user.password;
+
+    if (!hashedPassword) {
+      return res.status(401).json({
+        error: "HashError",
+        message: "Invalid login",
+      });
+    }
+
+    //CHECK IF THE ENTERED PASWORD MATCHES WITH THE HASHED ONE FROM DB
+    const isPasswordValid = await compare(enteredPassword, hashedPassword);
+    if (!isPasswordValid) {
+      return res.status(401).json({
+        error: "Unauthorized",
+        message: "Invalid login",
+      });
+    }
 
     //DELETE USER
     const userToDelete = await User.deleteOne({ _id: userId });
